Re-organize the fixture filler helper tools
