+++ docs/api/eth/api.tools.rst
@@ -1,0 +1,10 @@
+Tools
+=====
+
+
+.. toctree::
+   :maxdepth: 4
+   :name: toc-eth-api-tools
+   :caption: Tools
+
+   tools/api.tools.fillers
